+++ semana11/projeto/lista-de-tarefas/src/components/hooks/useInputValue.js
@@ -1,0 +1,15 @@
+import { useState } from "react";
+
+const useInput = initialState => {
+  const [inputValue, updateValue] = useState(initialState);
+
+  const handleUpdateValue = event => {
+    const { value } = event.target;
+
+    updateValue(value);
+  };
+
+  return [inputValue, handleUpdateValue];
+};
+
+export default useInput;
